+++ seihou-cli/test/Main.hs
@@ -4,6 +4,7 @@ import Seihou.CLI.BrowseFormatSpec qualified as BrowseFormatSpec
 import Seihou.CLI.DiffSpec qualified as DiffSpec
 import Seihou.CLI.InitSpec qualified as InitSpec
 import Seihou.CLI.ListSpec qualified as ListSpec
+import Seihou.CLI.SavePromptedSpec qualified as SavePromptedSpec
 import Seihou.CLI.UpgradeSpec qualified as UpgradeSpec
 import Seihou.FzfSpec qualified as FzfSpec
 import Test.Tasty
@@ -16,6 +17,7 @@ main = do
         DiffSpec.tests,
         InitSpec.tests,
         ListSpec.tests,
+        SavePromptedSpec.tests,
         UpgradeSpec.tests,
         FzfSpec.tests
       ]